refactor: move executemsg implementations around
