+++ contracts/liquidity_hub/vault-manager/src/vault/mod.rs
@@ -1,1 +1,1 @@
-
+pub mod commands;
