+++ packages/white-whale/src/lib.rs
@@ -4,9 +4,11 @@ pub mod constants;
 pub mod fee;
 pub mod fee_collector;
 pub mod fee_distributor;
+pub mod lp_common;
 pub mod migrate_guards;
 pub mod pool_network;
 pub mod token_factory;
+pub mod tokenfactory;
 pub mod traits;
 pub mod vault_manager;
 pub mod vault_network;
