+++ packages/white-whale/src/tokenfactory/mod.rs
@@ -1,0 +1,1 @@
+pub mod common;
